[refactor] remove EncryptionScheme
